+++ mtg_proxy_printer/decklist_parser/common.py
@@ -12,7 +12,7 @@
 
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
-
+import collections
 from abc import abstractmethod
 import typing
 
@@ -47,10 +47,36 @@ class ParserBase(QObject):
                 "prefer-already-downloaded"
             )
 
-    @abstractmethod
     def parse_deck(self, deck: str,
                    print_guessing: bool,
-                   print_guessing_prefer_already_downloaded: bool) -> ParsedDeck:
+                   print_guessing_prefer_already_downloaded: bool,
+                   language_override: str = None) -> ParsedDeck:
+
+        # Implementation note: If a language is given, force print_guessing_prefer_already_downloaded to False,
+        # Because it would operate on the cards in the source language. The card choice gets overwritten by the
+        # translation step, so performs unnecessary work that gets thrown away anyways.
+        self.print_guessing_prefer_already_downloaded = print_guessing_prefer_already_downloaded \
+            if language_override is None else False
+        parsed_deck, unmatched_lines = self.parse_deck_without_translation(deck, print_guessing)
+        self.print_guessing_prefer_already_downloaded = print_guessing_prefer_already_downloaded
+
+        translated_deck: typing.Counter[Card] = collections.Counter()
+        for card, count in parsed_deck.items():
+            if self.print_guessing_prefer_already_downloaded and \
+                    (all_printings := self.card_db.find_all_translated_printings(card, language_override)):
+                # Choose any printing, based on what is already downloaded. …
+                translated_card = self._determine_best_match(all_printings)
+                # … But if no already downloaded image is found, prefer accurate translations over random selections.
+                if translated_card is all_printings[0]:
+                    translated_card = self.card_db.translate_card(card, language_override)
+            else:
+                translated_card = self.card_db.translate_card(card, language_override)
+            translated_deck[translated_card] = count
+        return translated_deck, unmatched_lines
+
+    @abstractmethod
+    def parse_deck_without_translation(self, deck: str,
+                                       print_guessing: bool) -> ParsedDeck:
         """
         Parse the given deck.
 
@@ -89,6 +115,14 @@ class ParserBase(QObject):
                 f"Matching using language, set code and collector number. Found {len(possible_matches)} matches."
             )
             return self._determine_best_match(possible_matches)
+        if card_data.name and card_data.set_code and (
+                possible_matches := self.card_db.get_cards_from_data(CardIdentificationData(
+                    card_data.language, card_data.name, card_data.set_code
+                ))):
+            logger.debug(
+                f"Matching using language, card name and set code. Found {len(possible_matches)} matches."
+            )
+            return self._determine_best_match(possible_matches)
         if card_data.name and (
                 possible_matches := self.card_db.get_cards_from_data(CardIdentificationData(
                     card_data.language, card_data.name
@@ -112,4 +146,3 @@ class ParserBase(QObject):
         if self.add_opposing_face and (opposing_face := self.card_db.get_opposing_face(card)) is not None:
             # Double-faced card
             deck[opposing_face] += count
-